Eric

+++ src/server/Node.java
@@ -7,11 +7,11 @@ public class Node {
 	private Point2D.Double position;
 	private Building building;
 
-	public Node(int x, int y) {
+	public Node(double x, double y) {
 		this.position = new Point2D.Double(x, y);
 	}
 
-	public Node(int x, int y, Building building) {
+	public Node(double x, double y, Building building) {
 		this.position = new Point2D.Double(x, y);
 		this.building = building;
 	}
